primer paso para separa modelo y manufacturer de aircraft 😎👌

+++ backend/src/main/java/com/dronetools/dronegestory/dto/AircraftDocumentationDTO.java
@@ -3,8 +3,8 @@ package com.dronetools.dronegestory.dto;
 import java.time.LocalDate;
 
 public record AircraftDocumentationDTO(
-        Integer id,
-        Integer aircraftId,
+        Long id,
+        Long aircraftId,
         String documentationType,
         String documentationName,
         LocalDate expireDate,
