+++ backend/src/main/java/com/dronetools/dronegestory/repository/AircraftDocumentationRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AircraftDocumentationRepository extends JpaRepository<AircraftDocumentation, Integer> {
-    List<AircraftDocumentation> findByAircraftId(Integer aircraftId);
-    void deleteByAircraftId(Integer aircraftId);
+public interface AircraftDocumentationRepository extends JpaRepository<AircraftDocumentation, Long> {
+    List<AircraftDocumentation> findByAircraftId(Long aircraftId);
+    void deleteByAircraftId(Long aircraftId);
 }
